Zadacha 1, Posti
