+++ src/test/kotlin/ru/netology/WallServiceTest.kt
@@ -1,8 +1,6 @@
 package ru.netology
 
-import Likes
 import Post
-import WallService
 import org.junit.Test
 
 import org.junit.Assert.*
@@ -17,18 +15,18 @@ class WallServiceTest {
 
     @Test
     fun add() {
-        val result = WallService.add((Post(7, 1, "Hello", true, true, 13, 2, 1, "answer", Likes(10))))
+        val result = WallService.add((Post(7, 1, "Hello", true, true, 13, 2, 1, "answer",10, null)))
         assertEquals(result.id, 1)
     }
 
     @Test
     fun shouldBeTrueWhileUpdate() {
 
-        WallService.add(Post(1, 1, "Hello", true, true, 13, 2, 1, "answer", Likes(10)))
-        WallService.add(Post(2, 5, "Nice", true, true, 20, 3, 1, "type", Likes(9)))
-        WallService.add(Post(4, 7, "Parrot", true, false, 13, 2, 1, "top", Likes(1)))
+        WallService.add(Post(1, 1, "Hello", true, true, 13, 2, 1, "answer", 10, null))
+        WallService.add(Post(2, 5, "Nice", true, true, 20, 3, 1, "type", 9, null))
+        WallService.add(Post(4, 7, "Parrot", true, false, 13, 2, 1, "top", 1, null))
 
-        val update = Post(2, 6, "Nice", true, true, 14, 7, 2, "type", Likes(10))
+        val update = Post(2, 6, "Nice", true, true, 14, 7, 2, "type", 10, null)
 
         val result = WallService.update(update)
 
@@ -37,11 +35,11 @@ class WallServiceTest {
 
     @Test
     fun shouldBeFalseWhileUpdate() {
-        WallService.add(Post(1, 1, "Hello", true, true, 13, 2, 1, "answer", Likes(10)))
-        WallService.add(Post(2, 5, "Nice", true, true, 20, 3, 1, "type", Likes(9)))
-        WallService.add(Post(4, 7, "Parrot", true, false, 13, 2, 1, "top", Likes(1)))
+        WallService.add(Post(1, 1, "Hello", true, true, 13, 2, 1, "answer", 10, null))
+        WallService.add(Post(2, 5, "Nice", true, true, 20, 3, 1, "type", 9, null))
+        WallService.add(Post(4, 7, "Parrot", true, false, 13, 2, 1, "top", 1, null))
 
-        val update = Post(8, 6, "Nice", true, true, 14, 7, 2, "type", Likes(10))
+        val update = Post(8, 6, "Nice", true, true, 14, 7, 2, "type", 10, null)
 
         val result = WallService.update(update)
